Use prompt card to guide recording
Note: the recycler view is a fixed size to enable swipping between
cards. Therefore, when the margin of a prompt increases, so too should
this fixed this. In the future, this would be dynamic, but wrapping
content leads (as expected) to show all recyclerview items.

Made record “preview” gone as there is no space to show on the UI
without making the card smaller.