Record location as last previous location
Yes, this is an absolute tragic of an implementation given both the
try/catch and use of GetLastKnownLocation, which obviously may not be
where they are. However, it is “good enough”.

From an implementation perspective,

However, the try/catch is justified in that the GetBestProvider
requires a null check if/else block, so why not catch em’ all?