Update endpoints to match server
Likewise, made several changes to the /projects/ endpoint on the
server, which now returns a dictionary of `public` and `personal` (i.e.
the ones you are a member of). This was to reflect member
projects/public on the website and future app, so for now I have joined
both lists to reduce refactoring of code at present.