Send FCM token to API on iOS
We set FCM token in ProjectController as it is the entry point (after login) and FCM token is generated when the app is installed. We only care about if/when the user logs in. If they logout, then their token is cleared, so must be reset.

- Also removed unused notification methods

References: 9644bf160b056983f5f870afb32a93b6b5079c1e